fix windows problem

+++ recipes/cnpy/all/test_package/example1.cpp
@@ -5,9 +5,9 @@
 #include<map>
 #include<string>
 
-const int Nx = 128;
-const int Ny = 64;
-const int Nz = 32;
+const size_t Nx = 128;
+const size_t Ny = 64;
+const size_t Nz = 32;
 
 int main()
 {
